refactor: split helpers into several helper files (#1954)
